dom/nodes/Node-appendChild-script-and-source-from-fragment.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and source from a fragment</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<video id="media"></video>
<script>
const happened = [];
const media = document.getElementById("media");
test(() => {
  const source = document.createElement("source");
  const script = document.createElement("script");
  script.textContent = `
    happened.push(media.networkState);
`;

  const df = document.createDocumentFragment();
  df.appendChild(script);
  df.appendChild(source);

  assert_array_equals(happened, []);
  media.appendChild(df);
  assert_array_equals(happened, [HTMLMediaElement.NETWORK_NO_SOURCE]);
});
</script>
